+++ src/parser/validator.rs
@@ -33,7 +33,9 @@ impl<'help, 'cmd> Validator<'help, 'cmd> {
 
             let o = &self.cmd[&a];
             let should_err = if let Some(v) = matcher.args.get(&o.id) {
-                v.all_val_groups_empty() && o.get_min_vals() != Some(0)
+                v.all_val_groups_empty()
+                    && (o.get_min_vals() != Some(0)
+                        && o.get_num_vals().map(|r| r.min_values()) != Some(0))
             } else {
                 true
             };
@@ -248,6 +250,8 @@ impl<'help, 'cmd> Validator<'help, 'cmd> {
 
     fn validate_arg_num_vals(&self, a: &Arg, ma: &MatchedArg) -> ClapResult<()> {
         debug!("Validator::validate_arg_num_vals");
+        let mut min_vals_zero = false;
+
         if let Some(expected) = a.num_vals {
             for (_i, group) in ma.vals().enumerate() {
                 let actual = group.len();
@@ -255,13 +259,43 @@ impl<'help, 'cmd> Validator<'help, 'cmd> {
                     "Validator::validate_arg_num_vals: group={}, num_vals={}, actual={}",
                     _i, expected, actual
                 );
-                if expected != actual {
-                    debug!("Validator::validate_arg_num_vals: Sending error WrongNumberOfValues");
-                    return Err(Error::wrong_number_of_values(
+                min_vals_zero |= expected.min_values() == 0;
+                if let Some(expected) = expected.num_values() {
+                    if expected != actual {
+                        debug!(
+                            "Validator::validate_arg_num_vals: Sending error WrongNumberOfValues"
+                        );
+                        return Err(Error::wrong_number_of_values(
+                            self.cmd,
+                            a.to_string(),
+                            expected,
+                            actual,
+                            Usage::new(self.cmd)
+                                .required(&self.required)
+                                .create_usage_with_title(&[]),
+                        ));
+                    }
+                } else if actual < expected.min_values() {
+                    return Err(Error::too_few_values(
                         self.cmd,
                         a.to_string(),
-                        expected,
+                        expected.min_values(),
                         actual,
+                        Usage::new(self.cmd)
+                            .required(&self.required)
+                            .create_usage_with_title(&[]),
+                    ));
+                } else if expected.max_values() < actual {
+                    debug!("Validator::validate_arg_num_vals: Sending error TooManyValues");
+                    return Err(Error::too_many_values(
+                        self.cmd,
+                        ma.raw_vals_flatten()
+                            .last()
+                            .expect(INTERNAL_ERROR_MSG)
+                            .to_str()
+                            .expect(INVALID_UTF8)
+                            .to_string(),
+                        a.to_string(),
                         Usage::new(self.cmd)
                             .required(&self.required)
                             .create_usage_with_title(&[]),
@@ -288,7 +322,7 @@ impl<'help, 'cmd> Validator<'help, 'cmd> {
                 ));
             }
         }
-        let min_vals_zero = if let Some(num) = a.get_min_vals() {
+        if let Some(num) = a.get_min_vals() {
             debug!("Validator::validate_arg_num_vals: min_vals set: {}", num);
             if ma.num_vals() < num && num != 0 {
                 debug!("Validator::validate_arg_num_vals: Sending error TooFewValues");
@@ -302,9 +336,7 @@ impl<'help, 'cmd> Validator<'help, 'cmd> {
                         .create_usage_with_title(&[]),
                 ));
             }
-            num == 0
-        } else {
-            false
+            min_vals_zero |= num == 0;
         };
         // Issue 665 (https://github.com/clap-rs/clap/issues/665)
         // Issue 1105 (https://github.com/clap-rs/clap/issues/1105)